feat(multi-layers): Implement multi-layer support
This is a generalization of the existing concept of tri-layer support
that's already well known. Essentially, a multi-layer configuration
activates a particular layer (the "then-layer") when one or more other
layers (the "if-layers") are activated.

This is commonly used on ortho keyboards to activate a third "adjust"
layer while the primary two layers ("lower" and "raise") are active.